Wait for Turbolinks. See description.
Now that Turbolinks is in effect, clicking a link on js:true tests, will require the test to
wait for 'some' change to appear on page. Interacting with the page before Turbolinks is done
updating can cause such interactions to be overridden. In this case, edits to the textarea
were being done before Turbolinks loaded textarea data from the server.