Assign application fee refund to eligible applicant
This changes the target of the refund from team lead to one of the
applicants if the team lead does not have to pay upfront. It picks the
applicant by sorting them by name and assigning it to the first one
who has regular fee. The view has also been altered to sort applicants
by name.