Add basic cloning function to file picker
WIP: This clones the file input to hidden form and sets correct
attributes to ready it for upload, but leaves the original picker's
label element listener broken. It will have to be extracted into a
different component and handled. The root component doesn't handle
the addition of the new file yet, either.