Extract file picker to its own component
This extracts the file picker component and its jQuery watcher to
update the label into a single component. The use of the `key`
attribute with the component also allows it to be regenerated after
its inner `input` element has been moved to the hidden form, which
keeps the `onChange` behavior consistent.