Use Bullet to optimize N+1 queries on admin dashboard
This also fixes the abnormally slow loading of the dashboard on
the development environment. Oddly enough, the slowdown was caused
by bullet itself - the request continued even after a response was
rendered by Rails, waiting for Bullet to print its analysis of the
page. This was fixed by moving a too-eagerly added 'includes', to as
late as possible in the code, thus removing the Bullet slowdown (and
complaints about an unneccessary use of 'includes').