Prerequisites are checked against startup targets
Prior to this change, the check for prerequisites completion mistakenly
used the founder / startup role of the target being checked to compute
the status of prerequisite targets. This removes that link, and forces
prerequisite status to be checked against all timeline events posted
for a startup - this enforces an assumption that all prerequisites will
be startup role targets.