Allow admin to mark connect requests as 'cancelled'
A new status 'cancelled' was introduced. As of now, the admin is
given the provision to simply mark confirmed requests as cancelled.
This should ideally be accompanied by an email to all attendees.
Note: Deletion of the associated Zoom meeting and Google calendar
event is to be manually done by the admin as of now. An alert
reminding the same is shown.