Switch to using sv.localhost in development env
The .dev TLD is owned by Google, and they added .dev to the pre-loaded
HSTS list of Chrome on its last update, forcing SSL on all routes when
using Chrome. This doesn't work well with webpack-dev-server, which speeds
up asset compilation in development - useful when working with React
components. The .localhost TLD, on the other hand, is reserved by IETF
for use-cases such as ours.