Start styling with (S)CSS modules and fix related issues
This commit marks the beginning of the use of CSS modules in React
components, and it fixes a number of issues that I encountered while
setting up the environement to allow modules and SCSS compilation:

1. babel-plugin-react-css-modules requires that SCSS be compiled
(parsed?) with PostCSS, whereas the default webpacker setup uses
node-sass. So postcss-scss had to be added and configured in the
.babelrc file specifically to allow babel-plugin-react-css-modules
to work.

2. node-sass is awful - I wish webpacker gets rid of it in favor
of postcss-scss ASAP (if that's possible) because it kept throwing
unrelated errors which it finally reported as a mismatch in node
environment. The only way I could get it to accept that I had the
required bindings installed was to add it as a development dependency.
This is not ideal, but hey, c'est la vie.