Update default meta description on all pages
This introduces a new helper to set meta description (and
og:description) for all pages, which is used directly in layout files.
This change allows meta description to be set for each action from
en.yml, or override it from the action if it the content of the
description is dynamic.