Avoid duplicate call to load founder statuses [skip ci]
This refactors code to load founder statuses, removing the individual
route and instead using the target details route which was already
loading the required data (but in a parent component). This commit
changes the nature of the data passed by the server (cleans it up)
and passes it from the parent to the child component which displays
the deferred loaded data.