WIP: Getting ready to move to Rails 5.2
Includes harmless rails changes + a couple of gem updates.

'simple_form' had to be upgraded to v4.0. The stale 'simple_form_fancy_uploads'
was removed. 'active_admin' had to be pointed to master for 5.2 support, but
it is missing a fix from the latest release. Need to wait for the next release.