Major version upgrade for a few gems, minor for all others
This updates major version for slim, data_migrate, chartkick &
chromedriver_helper.

The update to chromedriver_helper required the addition of an explicit
require statement in the rails_helper file because the gem no longer
introduces a 'chromedriver' executable, opting to use
'chromedriver_helper' as the main executable. The require is needed to
set correct chromedriver executable path for Selenium / Capybara.

The version for jwt has also been locked to ~> 1.5 to avoid a major
version upgrade to v2+, whose consequences I could not determine.