OAuth failure should redirect back to original domain
This adds a failure handling action to `OmniauthCallbacksController`
that redirects the user back to the OAuth origin, if such an origin
exists. It also replaces the `oauth_unknown` endpoint with a general
`oauth_error` endpoint that accepts the flash message to be displayed
at the origin domain. This allows the route to be used for both OAuth
failure, and when the authenticated email couldn't be mapped to a
known user.