Add a workaround for a ActiveStorage::Blob.service assignment bug
This adds a hacky workaround to rails_helper, to attempt to manually
assign the value for ActiveStorage::Blob.service. Without this custom
retry code, the first automatic attempt will fail, causing specs
involving file uploads to crash. This workaround should be removed in
the future, when it is no longer required.