White-label public-facing resources library
This commit includes multiple changes. Details follow:

- This includes a few behavioral changes - the library used to list
private resources in the current student's school if they were not
linked to any target. With the updated behavior, only public resources
in a school, and private resources linked to targets in the current
student's course is displayed.
- The link to visit the Resource#show page from each entry listed on
the index has been removed, since the only resource that needs the show
page are streamable ones, and they have a play button that takes the
user to #show.
- Inaccessible resources now 404 instead of redirecting to the index
with a flash message.
- ResourcesHelper has been removed in favor of
Resources::IndexPresenter.
- A 'coaches_index_subheading' string has been introduced to hold the
subheading text on the library index page.