Extract EditorDrawer and DisablingCover components
The new EditorDrawer component holds the stlying of the drawer, and
accepts whatever goes in the drawer as children. Similarly, the new
DisablingCover component accepts interior elements as children, and
when disabled with a prop, draws a translucent cover over it, disabling
access to the contents.

These two new components should be further extracted to top-level SA
so that it can be used everywhere possible.