Make both agreement editors fully functional
This commit includes a number of changes:

1. Introduces Ruby GraphQL code to handle updation of school strings.
2. Populates and styles the agreements editor to handle both privacy
policy and terms of use.
3. Updates the disabling cover component to accept an optional list of
classes that would be applied to the cover's container div.
4. Creates a new GraphqlErrorHandler module that contains a functor.
The handler module created via the functor holds code that is required
in the error-handling path for GraphQL-promise-based queries. The
functor has been used in the older links form as well as the new
agreements form introduced in this commit.