Reintroduce sanitation on the front-end
This replaces the marked library that was being used in the front-end
to parse markdown with markdown-it, and introduces motion-markdown-it
into the Gemfile, a Ruby port of the JS library. This allows both the
client and the server to maintain (nearly) identical behavior when
parsing markdown, which involves parsing the MD through markdown-it
(commonmark strict), and then sanitizing it (Rails sanitize in server,
and a newly added sanitize-html lib client-side). This saves us from
having to parse user MD inputs sever-side. Presence of CSP should
prevent users from messing with the client-side sanitation process.