Improve sending of excluded submission IDs on review dash
...when loading previously reviewed submissions. The previous approach
used the Rails convention for passing an array data via the URL, but
did so without handling the case for an empty array. This commit
handles the empty array by not passing the param and replaces the Rails
convention param with a simple comma-separated list; because the list
of IDs can get pretty huge.