Introduce client-side Rollbar
This adds a ListUtils.unsafeFind method that accepts a message which
will be posted to Rollbar as an error if the find fails. The Rollbar
library defaults to not handling uncaught exceptions and unhandled
promise failures - this is configurable using environment variables.

Speaking of which, this commit also introduces environment variable
parsing to webpack, so that called to `process.env.WHATEVER` will be
translated to the environment-specific value in webpack's output for
the browser.