Rework the flow around marking a student as exited
This updates the form to remove the clear_coaches flag (which was
effectively unused) and removes the use of TeamUpService when marking
a student as exited, opting to write custom code instead. Specs related
to marking students as exited have been expanded to cover both
single-student teams and multi-student teams.