Use a reducer to manage state in MarkdownEditor
This is required because there was requirement that changing the
preview state of the editor should also remove the last stored
command, if any. This sort of mixed state change is best handled by a
reducer.

The presence of a previous command, when preview is toggled was
causing the underlying DraftEditor component to process the command
request, even though it had already been processed by a previous
(unmounted) instance of DraftEditor.