Add student tag filter to course report prep service
The tag filter allows the course report to include only a select
group of students, allowing admins to easily compare and contrast
within a defined group.

This commit also cleans up the list of students and submissions to
avoid dropped-out students and those whose access to a course has
ended.