Block target editor submit when link_to_complete is blank
This adds server-side validation for checking if the value of
link_to_complete is a proper URL (when it is present), and reworks
the UI code for handling submit values depending on the method of
completion. The submit button is also disabled when switching to
'visit link to complete' as a method of completion, if the field
is empty at that point.