Compare against latest value, not the original value
This fixes a logical mistake in making the decision to trigger the
callback supplied by the parent of MarkdownEditor. An incorrect
comparision was made between the _original_ value supplied to the
editor and the _new_ value to determine whether the callback should be
used.

The fix was to compare the _current_ value against the _new_ value,
and it fixes a bug that would cause the markdown editor in the school
admin target editor to refuse to update the content on single-character
edits. #67