Disable the submit button when the new coach form opens #84
This involved changing how the form computes disabled state for the
submit button, checking validity again instead of referring to the
hasThingError state used to display error messages under the input
fields. This commit also updates EmailUtils, converting the named
argument on the validity testing function to a regular one, and adding
the allowBlank argument to the UrlUtils validity testing function.