Upgrade most Javascript dependencies to latest
This was done using a yarn "rejuvenation" script:
https://gist.github.com/pftg/fa8fe4ca2bb4638fbd19324376487f42

This avoid CVE-2019-10747 in the set-value package.

I was also forced to fork the PrismJS library to fix a bug related
to improper use of the ternary operator in one of its plugins. I've
submitted a PR for that: https://github.com/PrismJS/prism/pull/2167