Only report dirty content block status when absolutely required
This improves performance of typing / editing anything about content
blocks in the new content editor massively. The earlier implementation
reported dirty status to the parent component for every single change
regardless of whether the dirty state had changed. This is fixed by
keeping a local state of the dirty status and reporting to parent only
if the the new state doesn't match the local state.

This commit also fixes a bad commenting-out of the previewing
functionality of the new markdown editor.