Create new Link component to properly manage routed links
For #211. This provides a standard pattern to follow when a link is
required that pushes a new route using the ReasonReact router. This
component has been set up to intercept Command / Control key mouse
clicks to properly open those requests in a new tab, and provides a
built-in confirmation function.