Use the Yarn workspaces feature to hoist monorepo packages
This commit resolves issues with the ReasonML compilation process
related to the non-availability of dependencies that were mentioned in
a sub-package's bsconfig.json file. The workspaces feature resolves
this by treating sub-packages as dependencies, while being aware that
the code is to be loaded from a sub-directory. This makes the
development process of libraries in the app/javascript/packages folder
much easier.