Fix a mistake in how submissions were selected for teams export
The comparison between a submission's founders and a team's current
founders was using a simplistic array compare, which was causing
the loop to reject many valid submissions. This is difficult to add
a test for, so I've issued the fix without an exact test that covers
this scenario. I've also added some optimizations on the selection
of submissions over which the report preparation service iterates.