Fix #322 by properly throwing error for unexpected type
The bug was caused by an improperly written Json decode method. All
such decode methods are supposed to perform a run-time typecheck and
throw a `DecodeError` if the supplied type is unexpected. This is what
allows the `optional()` decoder to catch and convert these values to
`None`. Implementing this type-check allows the custom
`DateFns.decodeISO` function to work in combination with the `optional`
decoder.