Add filtering of students by presence (or absence) of coach notes (#324)
In addition to the new filter, this adds clarification as to why students who have only have archived notes show up when the "has notes" filter is selected. Also reloads teams and distribution when when "does not have notes" filter is active and a student gets a new note from the coach.