Load IDs of prerequisite targets directly from the join table
Since the curriculum editor now automatically prevents draft and
archived targets from being prerequisites to other targets, there
is no need to filter the list of prerequisite target IDs passed to the
UI; the IDs stored in the join table can be safely used for this
purpose.

This commit also removes a spec that checked for UI behavior in the
invalid state of a target having a draft prerequisite target.