Allow setting draft targets as prerequisites to others (#569)
Prior to this change, setting a draft target as a prerequisite to a
live target would cause the student's curriculum UI to crash. This
change makes it so that such draft prerequisites are simply ignored
by the UI and the server. This commit also upgrades a lot of
unrelated ReScript code to the latest standards.