Enable translations fallback (#825)
Translations fall back to `I18N_DEFAULT_LOCALE`, or `en` if the
environment variable is missing.

While we don't allow the user to select their locale (yet), we do have a
column in the database that's meant to enable this. There's no reason
to not use this value right now. At least it allows testing to make sure
that locales are working as they should.