Attempt to fix build failure of developer docs
This follows instructions from a Docusaurus Github issue to try and
avoid a build failure caused by incompatibility of the alpha version
of Docusaurus v2 that we use with yarn.

This issue is fixed in the latest (beta) version of Docusaurus, but
its unusable because of its dependency on Webpack v5, which we do not
support at the moment.

See also: https://github.com/facebook/docusaurus/issues/3515#issuecomment-707145367