Add a temporary fix to failing calendar spec
Confirmation dialog check is removed for calendar event deletion
as UJS is missing in Vite test builds. Once we have a server rendered
layout for school admin pages we can have a custom entry point specific
to the page to handle this without relying on UJS and maybe even consider
using Turbo for server rendered pages.