Add :en to fallbacks when en isn't the default locale
Also set "en" as the frontend's `defaultLocale`, since it appears that
the only purpose of `defaultLocale` for `i18n-js` is to act as a
fallback for missing translations. The locale used in the frontend is
anyway dictated by the locale decided on by the server, which takes
default locale and user's choice (if any) into consideration.